Crate sid

Structs

Id

IdFreeList

A linked list stored in contiguous memory allowing random access through ids.

IdRange

IdSlice

IdVec

Similar to Vec except that it is indexed using an Id rather than an usize index. if the stored type implements Default, IdVec can also use the set(…) method which can grow the vector to accomodate for the requested id.

MutIdSlice

NoneAsNullId

Use this as the NullId type parameter of the freelist if the id type is an option, to use Option::None as the null id value.

ReverseIdRange

Traits

FromUsize

Identifier

IntegerHandle

NullId

A trait that defines how to choose the null (or invalid) Id.

ToUsize